separate out the signature registry tests because the small limit was interfering with benchmark tests
