+++ pallets/msa/src/mock.rs
@@ -107,10 +107,10 @@ impl pallet_msa::Config for Test {
 	type MaxProviderNameSize = MaxProviderNameSize;
 	type SchemaValidator = Schemas;
 	type MortalityWindowSize = ConstU32<100>;
-	type MaxSignaturesPerBucket = ConstU32<20>;
+	type MaxSignaturesPerBucket = ConstU32<4000>;
 	type NumberOfBuckets = ConstU32<2>;
 	/// This MUST ALWAYS be MaxSignaturesPerBucket * NumberOfBuckets.
-	type MaxSignaturesStored = ConstU32<40>;
+	type MaxSignaturesStored = ConstU32<8000>;
 }
 
 pub fn new_test_ext() -> sp_io::TestExternalities {
